allow specifying multiple pairs of files to mergepe.
also specify a comment for each pair e.g. a read-group so, e.g.:

cat <<EOF > hdr.txt
@RG	ID:l1	SM:sample	PL:ILLUMINA	PU:sample	LB:lib
@RG	ID:l2	SM:sample	PL:ILLUMINA	PU:sample	LB:lib
EOF

./seqtk mergepe lane1_{1,2}.fq 'RG:Z:l1' lane2_{1,2}.fq 'RG:Z:l2' \
   | bwa mem -C -H hdr.txt -p $refbase /dev/stdin \
   | samblaster \
   | samtools sort ...

This shows how we can use this to have a single stream of data from
bwa even though we have multiple lanes. In this case, we still want
to call duplicates (with samblaster) together because they are they
same library.